refactor: centralize user and organization caching logic within proxy utility and service layers
